Refactor timing and event handling in window system
Updated TimeEventArgs to use double for time properties.
Enhanced IWindow with new properties and methods.
Refactored SdlWindow to manage timing with Stopwatches.
Renamed HandleEvents to DoEvents across the codebase.
Introduced parallel loops for updates and renders in WindowManager.

+++ Source/Graphics/Graphics.Windowing/Events/TimeEventArgs.cs
@@ -1,8 +1,8 @@
 ﻿namespace Graphics.Windowing.Events;
 
-public class TimeEventArgs(float deltaTime, float totalTime) : EventArgs
+public class TimeEventArgs(double deltaTime, double totalTime) : EventArgs
 {
-    public float DeltaTime { get; } = deltaTime;
+    public double DeltaTime { get; } = deltaTime;
 
-    public float TotalTime { get; } = totalTime;
+    public double TotalTime { get; } = totalTime;
 }
